+++ src/euroeval/prompt_templates/summarization.py
@@ -1,7 +1,7 @@
 """Templates for the Summarization task."""
 
 from ..data_models import PromptConfig
-from ..languages import DA, DE, EN, ES, FR, IS, IT, NB, NL, NN, NO, SV
+from ..languages import DA, DE, EN, ES, FI, FR, IS, IT, NB, NL, NN, NO, SV
 
 # TODO: Missing Faroese
 SUMM_TEMPLATES = {
@@ -34,6 +34,14 @@ SUMM_TEMPLATES = {
         default_prompt_template="Documento: {text}\nResumen: {target_text}",
         default_instruction_prompt="Documento: {text}\n\nEscriba un resumen del "
         "documento anterior.",
+        default_prompt_label_mapping=dict(),
+    ),
+    FI: PromptConfig(
+        default_prompt_prefix="Seuraavassa on artikkeleita ja niihin liittyviä "
+        "tiivistelmiä.",
+        default_prompt_template="Uutisartikkeli: {text}\nTiivistelmä: {target_text}",
+        default_instruction_prompt="Uutisartikkeli: {text}\n\nKirjoita tiivistelmä "
+        "yllä olevasta artikkelista.",
         default_prompt_label_mapping=dict(),
     ),
     FR: PromptConfig(